Refactor leaf_aabb into leaf_masks
