+++ deadleaves/acceleration/quadtree.py
@@ -8,86 +8,9 @@ frontier of remaining gaps is visited.
 """
 
 from __future__ import annotations
-import math
 import torch
 
-__all__ = ["leaf_aabb", "CoverageQuadTree"]
-
-
-def leaf_aabb(
-    params: dict[str, torch.Tensor],
-    leaf_shape: str,
-) -> tuple[int, int, int, int]:
-    """Compute the axis-aligned bounding box of a leaf.
-
-    Args:
-        params:
-            Sampled leaf parameters (x_pos, y_pos, area / radius, …).
-        leaf_shape:
-            One of "circular", "ellipsoid", "rectangular", "polygon".
-
-    Returns:
-        (y_min, x_min, y_max, x_max) as ints, not yet clipped to canvas.
-    """
-    cx = float(params["x_pos"])
-    cy = float(params["y_pos"])
-
-    if leaf_shape == "circular":
-        keys = params.keys() if isinstance(params, dict) else params.index
-        if "radius" in keys:
-            r = float(params["radius"])
-        else:
-            r = math.sqrt(float(params["area"]) / math.pi)
-        return (
-            math.floor(cy - r),
-            math.floor(cx - r),
-            math.ceil(cy + r),
-            math.ceil(cx + r),
-        )
-
-    if leaf_shape in ("ellipsoid", "rectangular"):
-        area = float(params["area"])
-        aspect = float(params["aspect_ratio"])
-        theta = float(params["orientation"])
-
-        if leaf_shape == "ellipsoid":
-            a = math.sqrt(area * aspect / math.pi)
-            b = math.sqrt(area / (math.pi * aspect))
-            # Ellipse: bounding half-widths use Pythagorean formula
-            cos_t = math.cos(theta)
-            sin_t = math.sin(theta)
-            hx = math.sqrt((a * cos_t) ** 2 + (b * sin_t) ** 2)
-            hy = math.sqrt((a * sin_t) ** 2 + (b * cos_t) ** 2)
-        else:  # rectangular
-            h = math.sqrt(area / aspect)
-            w = h * aspect
-            a = w / 2
-            b = h / 2
-            # Rectangle: bounding half-widths use absolute-value sum
-            abs_cos = abs(math.cos(theta))
-            abs_sin = abs(math.sin(theta))
-            hx = a * abs_cos + b * abs_sin
-            hy = a * abs_sin + b * abs_cos
-        return (
-            math.floor(cy - hy),
-            math.floor(cx - hx),
-            math.ceil(cy + hy),
-            math.ceil(cx + hx),
-        )
-
-    if leaf_shape == "polygon":
-        area = float(params["area"])
-        n_v = int(params["n_vertices"])
-        r = math.sqrt(2 * area / (n_v * math.sin(2 * math.pi / n_v)))
-        return (
-            math.floor(cy - r),
-            math.floor(cx - r),
-            math.ceil(cy + r),
-            math.ceil(cx + r),
-        )
-
-    # Fallback — unknown shape, return infinite box (no acceleration)
-    return (-(10**9), -(10**9), 10**9, 10**9)
+__all__ = ["CoverageQuadTree"]
 
 
 class Node:
